Made the read, write index a uint16_t instead of uint8_t because it could only hold 256 readings instead of the ~3K the ram is capable of

+++ Src/ATmega328p/ATmega328p.ino
@@ -100,8 +100,8 @@ unsigned long lastRainBucketReadingMillis = 0;
 
 WeatherReadingAccumulator sampleAccumulator;
 
-uint8_t weatherReadingWriteIndex = 0;
-uint8_t weatherReadingReadIndex = 0;
+uint16_t weatherReadingWriteIndex = 0;
+uint16_t weatherReadingReadIndex = 0;
 
 #define ESP_STATE_SLEEP 0
 #define ESP_STATE_IDLE 1
